feat: add id_municipio parameter to ProposicoesCrawler constructor
This commit modifies the `ProposicoesCrawler` class by adding a new `id_municipio` parameter to the constructor. This parameter allows the crawler to fetch data specific to a municipality.

The code changes include:
- Adding a new parameter `id_municipio` to the `__init__` method of `ProposicoesCrawler`
- Assigning the value of `id_municipio` to the `self.id_municipio` attribute

This feature improves the flexibility of the `ProposicoesCrawler` class by allowing it to fetch data for different municipalities based on the provided `id_municipio`.