feat: improve Dockerfile and add Swagger documentation link
The Dockerfile has been updated to copy all files instead of just the `app` directory. This ensures that all necessary files are included in the container. Additionally, the import statement for the `app` module has been fixed in the CMD command.

This commit also adds a Swagger documentation link to the readme file.

feat: add endpoint to fetch proposicao tipos

A new endpoint `/proposicao/tipos` has been added to the FastAPI application. This endpoint returns a list of proposicao tipos with their corresponding IDs and names.

feat: add id_municipio parameter to ProposicoesCrawler constructor

The `ProposicoesCrawler` class has been modified to include a new `id_municipio` parameter in the constructor. This allows the crawler to fetch data specific to a municipality.

feat: add endpoint to generate Excel file with data

A new endpoint `/generate-excel` has been added to the FastAPI application. This endpoint accepts parameters `tipo`, `data_inicio`, and `data_final` to generate an Excel file with data.

chore: projeto inicial

This is the initial commit of the project.